Added lots of functionality and connected the database
