+++ nbproject/project.properties
@@ -30,6 +30,7 @@ dist.jar=${dist.dir}/MyAttendance.jar
 dist.javadoc.dir=${dist.dir}/javadoc
 endorsed.classpath=
 excludes=
+file.reference.joda-time-2.9.7-dist.zip=/Users/Kristoffers/Downloads/joda-time-2.9.7-dist.zip
 file.reference.sqljdbc42.jar=..\\JDBCTest\\sqljdbc42.jar
 includes=**
 # Non-JavaFX jar file creation is deactivated in JavaFX 2.0+ projects
@@ -37,7 +38,8 @@ jar.archive.disabled=true
 jar.compress=false
 javac.classpath=\
     ${javafx.classpath.extension}:\
-    ${file.reference.sqljdbc42.jar}
+    ${file.reference.sqljdbc42.jar}:\
+    ${file.reference.joda-time-2.9.7-dist.zip}
 # Space-separated list of extra javac options
 javac.compilerargs=
 javac.deprecation=false
