+++ src/myattendance/BLL/DateManager.java
@@ -1,0 +1,27 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package myattendance.BLL;
+
+import java.time.temporal.ChronoUnit;
+import java.util.Date;
+import myattendance.DAL.DALFacade;
+
+
+/**
+ *
+ * @author Kristoffers
+ */
+public class DateManager
+{
+    DALFacade dalFacade = DALFacade.getInstance();
+    
+    
+    public int daysDiff(String startDate, String currentDate)
+    {
+        String startDate = 
+    }
+    
+}
