+++ src/myattendance/DAL/DatabaseAccess.java
@@ -3,10 +3,13 @@ package myattendance.DAL;
 import com.microsoft.sqlserver.jdbc.SQLServerDataSource;
 import com.sun.rowset.CachedRowSetImpl;
 import java.sql.Connection;
+import java.sql.Date;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Statement;
+import java.util.logging.Level;
+import java.util.logging.Logger;
 import myattendance.BE.Student;
 
 /**
@@ -81,4 +84,6 @@ public class DatabaseAccess
         }
 
     }
+
+
 }
