+++ src/myattendance/GUI/Controller/StudentMainOverviewController.java
@@ -76,14 +76,14 @@ public class StudentMainOverviewController implements Initializable
     private void updateView()
     {
         ObservableList<PieChart.Data> pieChartData = FXCollections.observableArrayList();
-        pieChartData.add(new PieChart.Data("Absence", student.getAbsentClasses()));
-        pieChartData.add(new PieChart.Data("Presence", student.getPresentClasses()));
+        pieChartData.add(new PieChart.Data("Absence", student.getAbsentDates()));
+        pieChartData.add(new PieChart.Data("Presence", student.getPresentDates()));
 
         PieChart absenceChart = new PieChart(pieChartData);
         absenceChart.setTitle("Absence");
 
         Label absenceLabel = new Label();
-        absenceLabel.setText("Student Attendance: " + student.getPresentClasses() + "/" + Math.addExact(student.getAbsentClasses(), student.getPresentClasses()));
+        absenceLabel.setText("Student Attendance: " + student.getPresentDates() + "/" + Math.addExact(student.getAbsentDates(), student.getPresentDates()));
         
         vBoxMiddle.getChildren().add(absenceChart);
         vBoxMiddle.getChildren().add(absenceLabel);
