--- src/myattendance/GUI/View/StatisticAttendanceOverview.fxml
@@ -1,27 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.TableColumn?>
-<?import javafx.scene.control.TableView?>
-<?import javafx.scene.control.TextArea?>
-<?import javafx.scene.image.Image?>
-<?import javafx.scene.image.ImageView?>
-<?import javafx.scene.layout.AnchorPane?>
-
-<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/8.0.60" xmlns:fx="http://javafx.com/fxml/1" fx:controller="myattendance.GUI.Controller.StatisticAttendanceOverviewController">
-   <children>
-      <ImageView fitHeight="150.0" fitWidth="200.0" layoutX="14.0" layoutY="14.0" pickOnBounds="true" preserveRatio="true" AnchorPane.leftAnchor="10.0" AnchorPane.topAnchor="10.0">
-         <image>
-            <Image url="@Resource/logo.png" />
-         </image>
-      </ImageView>
-      <Button fx:id="overviewButton" layoutX="14.0" layoutY="65.0" mnemonicParsing="false" onAction="#handleHomepage" text="Homepage" AnchorPane.leftAnchor="15.0" AnchorPane.topAnchor="70.0" />
-      <TableView layoutX="15.0" layoutY="107.0" prefHeight="265.0" prefWidth="168.0">
-        <columns>
-          <TableColumn prefWidth="88.0" text="First Name" />
-          <TableColumn prefWidth="78.0" text="Last Name" />
-        </columns>
-      </TableView>
-      <TextArea layoutX="275.0" layoutY="70.0" prefHeight="308.0" prefWidth="305.0" text="This view is intended for later implementation.&#10;It should load inside the teachers homepage &#10;and work as a manual absence correction.&#10;" />
-   </children>
-</AnchorPane>
